refactor: wishlist card
