+++ src/component/index.js
@@ -11,3 +11,4 @@ export {Modal} from "./modal/Modal";
 export {AddressModal} from "./modal/AddressModal";
 export {Filters} from "./filters/Filters"
 export {Loader} from "./loader/Loader"
+export {WishlistCard} from "./card/WishlistCard"
